Support the new OpenSSL TLS_method
In the newest version, OpenSSL replaced `SSLv23_method` with
`TLS_method`, then used a macro to link the two. Since we can't
detect the macro, it was not being found.

This causes us to look for `TLS_method` and then use it as a
replacement for `SSLv23_method` if necessary.